Better concurrency, and multicast announcements
The foreachUpstream function in the server is now calling its designated
functions concurrently, each one on its own goroutine. This simplified
the code of the handleTables http handler, which had a bug and was being
run serial not concurrent.

A new feature has been introduced: the server can announce itself to
other servers via multicast so we can create chains of aggregations.
Although the mechanism is in place the parsing of http responses may
fail due to malformed json, because of the two or more levels of
aggregation. Will fix this when the time comes.